dev-lang/perl: Fix version in hardcoded config
It turns out the config.sh generated by perl contains the version
number in various contexts, so when we copy it in for the build many of
the output files end up in the wrong place.  Introduce a way to replace
variables in this file while copying it and use this to replace the
hardcoded version with ${PN}.

Another pass will be needed to ensure that all of the variables that can
be influenced by the environment are set correctly (for example the libc
version is hardcoded).